pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "climate-analysis"
version = "0.1.0"
description = "Tools for downloading and analyzing climate data"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
dependencies = [
    "aiohttp>=3.9.4",
    "cftime>=1.6.0",
    "dask[complete]>=2024.9.0",
    "fsspec>=2024.0.0",
    "gcsfs>=2024.6.1",
    "h5netcdf>=1.3.0",
    "intake>=2.0.0",
    "intake-esm>=2024.2.23",
    "intake-xarray>=0.7.0",
    "matplotlib>=3.8.0",
    "meteostat>=1.6.7",
    "netcdf4>=1.6.0",
    "numpy>=1.26.0",
    "pandas>=2.0.0",
    "pooch>=1.8.0",
    "pydap>=3.4.0",
    "pyyaml>=6.0.2",
    "requests>=2.31.0",
    "scipy>=1.11.0",
    "xarray>=2024.0.0",
    "zarr>=2.17.0",
]

[project.optional-dependencies]
notebooks = [
    "jupyter>=1.0.0",
    "jupyterlab>=4.2.5",
    "ipywidgets>=8.1.0",
]

[tool.uv]
# uv configuration

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.4.0",
]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = []

[tool.hatch.build.targets.wheel]
packages = ["climate_analysis", "climate_tools"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
